optimize IKS install

+++ iks/install.sh
@@ -14,31 +14,34 @@ elif [[ "$OS" == "linux" ]]; then
       fi
     else
       # Install for Jenkins alpine
-      apk add jq
+      if [ -z "$(which jq)" ]; then
+          apk add jq
+      fi
     fi
 else
     echo "Unsupported OS"
     exit 1
-
 fi
 
-wget https://clis.cloud.ibm.com/download/bluemix-cli/latest/linux64 -O ibmcloud-cli.tar.gz
-tar -xvzf ibmcloud-cli.tar.gz
-rm ibmcloud-cli.tar.gz
-Bluemix_CLI/install
-if [ $? -ne 0 ]; then
-    echo "IBM Cloud CLI installation failed, exiting with a failure"
-    exit 1;
-else
-    echo "IBM Cloud CLI version $(ibmcloud version) installed successfullly."
-fi
+if [ -z "$(which ibmcloud)" ]; then
+    wget https://clis.cloud.ibm.com/download/bluemix-cli/latest/linux64 -O ibmcloud-cli.tar.gz
+    tar -xvzf ibmcloud-cli.tar.gz
+    rm ibmcloud-cli.tar.gz
+    Bluemix_CLI/install
+    if [ $? -ne 0 ]; then
+      echo "IBM Cloud CLI installation failed, exiting with a failure"
+      exit 1;
+    else
+      echo "IBM Cloud CLI version $(ibmcloud version) installed successfullly."
+    fi
 
-ibmcloud version
-curl --progress-bar -LO https://storage.googleapis.com/kubernetes-release/release/$(curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt)/bin/linux/amd64/kubectl
-ibmcloud plugin install container-service -f
-if [ $? -ne 0 ]; then
-    echo "IBM Cloud CLI kubernetes plugin installation failed, exiting with a failure"
-    exit 1;
-else
-    echo "IBM Cloud CLI kubernetes pluing version installed successfullly."
+    ibmcloud version
+    curl --progress-bar -LO https://storage.googleapis.com/kubernetes-release/release/$(curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt)/bin/linux/amd64/kubectl
+    ibmcloud plugin install container-service -f
+    if [ $? -ne 0 ]; then
+      echo "IBM Cloud CLI kubernetes plugin installation failed, exiting with a failure"
+      exit 1;
+    else
+      echo "IBM Cloud CLI kubernetes pluing version installed successfullly."
+    fi
 fi
